pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linear-mcp"
version = "0.1.0"
description = "Linear MCP for Claude"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.24.0",
    "mcp>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
linear-mcp = "linear_mcp.main:main"
